moved db from src to database folder
