+++ backend_database/src/plugins/databasePlugin.ts
@@ -2,7 +2,7 @@ import sqlite3 from "sqlite3";
 import fp from "fastify-plugin";
 import { FastifyInstance } from "fastify";
 import { config } from "../config.ts";
-import { runMigrations } from "../database/migrator.ts";
+import { runMigrations } from "../db_migrator/migrator.ts";
 
 interface DatabaseOptions {
   path?: string;
